// src/common/utils/role.enum.ts
export enum RoleBase {
    ADMIN = 'ADMIN',
    USER_DEPT = 'USER_DEPT',
    ADMIN_DEPT = 'ADMIN_DEPT',
    INVENTORY_COMMITTEE_HEAD = 'INVENTORY_COMMITTEE_HEAD',
    INVENTORY_COMMITTEE_MEMBER = 'INVENTORY_COMMITTEE_MEMBER',
    INVENTORY_COMMITTEE_SECRETARY = 'INVENTORY_COMMITTEE_SECRETARY',
    
}